Refactor how COPs are tracked and deparsed
Added a set of methods to track the current COP:
    _cop_stack(), _cur_cop(), _set_cur_cop(), _enter_scope(), leave_scope()
A stack of current COPs is stored on the root of the OpTree.  Scope-like ops
push a new entry onto the stack before deparsing their contents, and pop it
off before returning.

COPs now have contents to deparse.  If they're not the first COP of the
current scope, they deparse to a semicolon.  They also track the number
of empty lines between the last COP and itself, and insert the appropriate
number of newlines.