Add hooks for finding OpTree objs given their B::OP
This lets us provide methods for next() sibling(), etc

The references aren't stored right on the OpTree objects.  This helps when
debugging problems.  If the refs were stored on the objects, then putting one
of the OpTree objets in the debugger's watch list could generate an infinite
loop, as these refs could point back to prior points in the parse tree.