Fix at_end() detection for perl 5.8
This test succeedes on all Perls except 5.8(.9).  Apparently the $AT_END
lexical variable goes out of scope before the code that checks it can run.
Changing it to a package-global seems to fix the problem, and also works
on the other Perls.

I also changed the regex matching END as a subname to use substr(), which
is faster.