Turn on source-code saving debugging flag sooner
Seems that on threaded perls, the elements in main:: that store the loaded
file source aren't populated unless the 0x400 debugging flag is on when the
files is loaded.  Non-threaded perls don't behave that way.

This was causing the tests covering is_loaded() and loaded_files() to fail on
threaded perls: 10-loaded-files.t and 11-postponed.t